reorganized ajax file
