+++ app/assets/stylesheets/pages/_dashboard.scss
@@ -57,7 +57,7 @@ h2 {
 }
 
 #button {
-  position: absolute;
+  position: fixed;
   right: 20px;
   bottom: 20px;
 }
